decor and puzzle inception
i added some decoration around the lakeside labyrinth to make it less bland. i also invented and worked on the mechanics for a puzzle involving geysers and switches. only the first one is done and its simple, but i will add them in across the entire maze in a later commit.