moved p2's story cutscene from verdin inn f2 to lookout point
a ton of stuff was changed as a result of this. a whole new cutscene was devised for verdin inn f2, featuring a new track (theme_mage.ogg). the mage's theme is also present in the initial bandito cutscene now. additionally, player 3's story cutscene is inaccessible at the moment but will be re-added in a future commit to a different area.