great pyramid spike retraction bug fix
previously, spikes would be retracted by the buttons, and while the player could always walk over them, they would still show as being up when leaving and re-entering the map. this was due to the <Save Event Locations> notetag on the map, which has been removed. now all boulders have been given the <Save Event Location>, similarly to the ruined fort, which is why the ruined fort's spikes worked before but the great pyramid's didn't.