SP value draw function no longer crashes game when opening main menu
it was crashing because this._actor.barrierPoints() and this._actor.mmp don't exist unless you're in battle, and they were used to evaluate the hp text drawing. i added a $gameParty.inBattle() eval before the two so that it will stop the whole eval before the two invalid evals if the party is not in battle.