phoenix feather replica now has sell price of 1
yes, this is for a reason. it's the same logic as the essences, giving it a very low sell value makes it apparent that you're supposed to keep it and not sell it. i would just make it unable to be sold, but i didn't just for the sake of consistency with other non-key items.