Warden and VotFG now use Engulfing Darkness instead of Darkfire
out of the desire to make Darkfire a Laeryidyean-exclusive signature move, i made a copy of Darkfire, changed the name, then adjusted the ai of the Warden and VotFG to use it instead of Darkfire. other than the name, animation, and lack of an mdf debuff, it's the same skill.