Changes to Solus Map and Overworld Solus
I changed the map for Solus town to add streetlamps that have a wide radius for light to allow more lighting in Solus during light. I had to change a character's pathing to go around one of the lamps. I also changed NPC dialogue so that when it is night, they have appropriate dialogue such as "good night" instead of "nice day". I also changed the overworld so it is also affected by the terrax lighting, and added a lighting effect around solus.