added the Seed Maker which will be the new way of getting seeds
we're not there yet, but the farmer will give the player 5 of each type of seed to start with. i'm not sure how i will make seeds renewable if the player accidentally uses all of their crops, maybe i'll make beet boxers have a small chance of dropping any seed or something.