removed proficiencies
it was an interesting system and was definitely coming along, but at the end of the day it's still me trying to force a Stardew Valley-esque feature into a game that just doesn't need it. the game's progression is mainly based on collecting powerful items and equipment, so progression of these specific activities will also follow the same route, as can already be seen by the various pickaxes/fishing rods and new farming items.