heavily nerfed the crit formula, again
p4 was just too strong, completely cheesing boss fights such as Dilong and the Thief Duo. this might make some fights a little too hard if people were using her as a crutch, so balance will be done as necessary.
this change also heavily decreases the amount of damage enemies do on crits, so hopefully this reduces the amount of one-shotting.