made Will's weapon much stronger
they've been in a weird spot in progression for a while, considering that they're not unlocked until after unlocking the forge, but this also unlocks the shovel that lets you go to telluria castle and buy new weapons. all they were useful for before was the Dandy Toad raid, but now that their stats are much closer to those of TCT weapons, they're useful as alternatives if the player would rather spend their money elsewhere.